ensures that user dont manipulate other user profile data by url routing.

+++ project.py
@@ -280,9 +280,11 @@ def new_team():
 @app.route("/team/<int:team_id>/edit/", methods=['GET', 'POST'])
 def edit_team(team_id):
     # return "This page will edit Team %s" % team_id
+    editedTeam = session.query(Team).filter_by(id=team_id).one()
     if 'username' not in login_session:
         return redirect('/login')
-    editedTeam = session.query(Team).filter_by(id=team_id).one()
+    if login_session['user_id'] != editedTeam.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to manipulate this Team data. Please create your own Team in order to do so.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['name']:
             editedTeam.name = request.form['name']
@@ -298,9 +300,11 @@ def edit_team(team_id):
 @app.route("/team/<int:team_id>/delete/", methods=['GET', 'POST'])
 def delete_team(team_id):
     # return "This page will delete Team %s" % team_id
+    deletedTeam = session.query(Team).filter_by(id=team_id).one()
     if 'username' not in login_session:
         return redirect('/login')
-    deletedTeam = session.query(Team).filter_by(id=team_id).one()
+    if login_session['user_id'] != deletedTeam.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to manipulate this Team data. Please create your own Team in order to do so.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(deletedTeam)
         session.commit()
@@ -328,9 +332,11 @@ def show_players(team_id):
 @app.route("/team/<int:team_id>/player/new/", methods=['POST', 'GET'])
 def new_player(team_id):
     # return "This page add new player in Team %s" % team_id
+    team = session.query(Team).filter_by(id=team_id).one()
     if 'username' not in login_session:
         return redirect('/login')
-    team = session.query(Team).filter_by(id=team_id).one()
+    if login_session['user_id'] != team.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to manipulate this Team data. Please create your own Team in order to do so.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         if request.form['name']:
             # step 9. add player creator user_id details while creating new player in team
@@ -359,10 +365,13 @@ def new_player(team_id):
 @app.route("/team/<int:team_id>/player/<int:player_id>/edit/", methods=['POST', 'GET'])
 def edit_player(team_id, player_id):
     # return "This page will edit player %s of Team %s" % (player_id, team_id)
-    if 'username' not in login_session:
-        return redirect('/login')
     team = session.query(Team).filter_by(id=team_id).one()
     editedPlayer = session.query(Player).filter_by(id=player_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if login_session['user_id'] != team.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to manipulate this Team data. Please create your own Team in order to do so.');}</script><body onload='myFunction()''>"
+
     if request.method == 'POST':
         if request.form['name']:
             editedPlayer.name = request.form['name']
@@ -397,10 +406,13 @@ def edit_player(team_id, player_id):
 @app.route("/team/<int:team_id>/player/<int:player_id>/delete/", methods=['POST', 'GET'])
 def delete_player(team_id, player_id):
     # return "This page will delete player %s of Team %s" % (player_id, team_id)
-    if 'username' not in login_session:
-        return redirect('/login')
     team = session.query(Team).filter_by(id=team_id).one()
     deletedPlayer = session.query(Player).filter_by(id=player_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if login_session['user_id'] != team.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to manipulate this Team data. Please create your own Team in order to do so.');}</script><body onload='myFunction()''>"
+
     if request.method == 'POST':
         session.delete(deletedPlayer)
         session.commit()
